requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
httpx



ruff
pytest